refactor: split main into modules and update module path
