+++ main.go
@@ -1,1133 +1,12 @@
 package main
 
 import (
-	"bufio"
-	"context"
-	"crypto/sha256"
-	"encoding/base64"
-	"errors"
 	"flag"
 	"fmt"
-	"io"
-	"io/fs"
-	"mime"
 	"os"
-	"path/filepath"
-	"regexp"
-	"runtime"
-	"strconv"
-	"strings"
-	"sync"
-	"time"
 
-	"github.com/bmatcuk/doublestar/v4"
-	"github.com/mark3labs/mcp-go/mcp"
 	"github.com/mark3labs/mcp-go/server"
 )
-
-// ---- Types ----
-
-type writeStrategy string
-
-const (
-	strategyOverwrite    writeStrategy = "overwrite"
-	strategyNoClobber    writeStrategy = "no_clobber"
-	strategyAppend       writeStrategy = "append"
-	strategyPrepend      writeStrategy = "prepend"
-	strategyReplaceRange writeStrategy = "replace_range"
-)
-
-type encodingKind string
-
-const (
-	encText   encodingKind = "text"
-	encBase64 encodingKind = "base64"
-)
-
-// ---- Helpers ----
-
-func detectMIME(name string, sample []byte) string {
-	if ext := filepath.Ext(name); ext != "" {
-		if mt := mime.TypeByExtension(ext); mt != "" {
-			return mt
-		}
-	}
-	if isText(sample) {
-		return "text/plain; charset=utf-8"
-	}
-	return "application/octet-stream"
-}
-
-func isText(b []byte) bool {
-	for _, c := range b {
-		if c == 9 || c == 10 || c == 13 {
-			continue
-		}
-		if c < 32 || c == 0x7f {
-			return false
-		}
-	}
-	return true
-}
-
-func sha256sum(b []byte) string {
-	s := sha256.Sum256(b)
-	return fmt.Sprintf("%x", s[:])
-}
-
-func ensureParent(path string) error {
-	dir := filepath.Dir(path)
-	return os.MkdirAll(dir, 0o755)
-}
-
-// trimUnderRoot returns p relative to root without a leading slash.
-// It normalizes separators and handles the case where root is "/".
-func trimUnderRoot(root, p string) string {
-	r := mustAbs(root)
-	r = strings.TrimSuffix(r, string(os.PathSeparator))
-	prefix := r + string(os.PathSeparator)
-	return strings.TrimPrefix(p, prefix)
-}
-
-func readWindow(path string, offset, max int) ([]byte, int64, bool, error) {
-	f, err := os.Open(path)
-	if err != nil {
-		return nil, 0, false, err
-	}
-	defer f.Close()
-	fi, err := f.Stat()
-	if err != nil {
-		return nil, 0, false, err
-	}
-	sz := fi.Size()
-	if offset < 0 {
-		offset = 0
-	}
-	if int64(offset) > sz {
-		return []byte{}, sz, true, nil
-	}
-	if _, err := f.Seek(int64(offset), io.SeekStart); err != nil {
-		return nil, sz, false, err
-	}
-	if max <= 0 {
-		max = defaultPeekMaxBytes
-	}
-	buf := make([]byte, max)
-	n, err := f.Read(buf)
-	if err != nil && !errors.Is(err, io.EOF) {
-		return nil, sz, false, err
-	}
-	buf = buf[:n]
-	return buf, sz, int64(offset+n) >= sz, nil
-}
-
-func parseMode(s string) (os.FileMode, error) {
-	if s == "" {
-		return 0o644, nil
-	}
-	if !strings.HasPrefix(s, "0") {
-		s = "0" + s
-	}
-	u, err := strconv.ParseUint(s, 0, 32)
-	if err != nil {
-		return 0, err
-	}
-	return os.FileMode(u), nil
-}
-
-// atomicWrite writes to a temp file then renames over target.
-func atomicWrite(target string, data []byte, mode os.FileMode) error {
-	dir := filepath.Dir(target)
-	tmp, err := os.CreateTemp(dir, ".mcpfs-*")
-	if err != nil {
-		return err
-	}
-	tmpName := tmp.Name()
-	defer func() { _ = os.Remove(tmpName) }()
-	if _, err := tmp.Write(data); err != nil {
-		tmp.Close()
-		return err
-	}
-	if err := tmp.Chmod(mode); err != nil {
-		tmp.Close()
-		return err
-	}
-	if err := tmp.Close(); err != nil {
-		return err
-	}
-	if err := os.Rename(tmpName, target); err != nil {
-		if runtime.GOOS == "windows" {
-			if removeErr := os.Remove(target); removeErr != nil && !os.IsNotExist(removeErr) {
-				return removeErr
-			}
-			return os.Rename(tmpName, target)
-		}
-		return err
-	}
-	return nil
-}
-
-// acquireLock creates a best-effort advisory lock using a sibling .lock file.
-// It evicts lock files older than ten minutes.
-func acquireLock(path string, timeout time.Duration) (release func(), err error) {
-	lock := path + ".lock"
-	deadline := time.Now().Add(timeout)
-	for {
-		f, err := os.OpenFile(lock, os.O_CREATE|os.O_EXCL|os.O_WRONLY, 0o644)
-		if err == nil {
-			_, _ = fmt.Fprintf(f, "%d\n", os.Getpid())
-			_ = f.Close()
-			return func() { _ = os.Remove(lock) }, nil
-		}
-		if !errors.Is(err, os.ErrExist) {
-			return nil, err
-		}
-		if info, statErr := os.Stat(lock); statErr == nil {
-			if time.Since(info.ModTime()) > 10*time.Minute {
-				_ = os.Remove(lock)
-				continue
-			}
-		}
-		if time.Now().After(deadline) {
-			return nil, fmt.Errorf("lock timeout: %s", path)
-		}
-		time.Sleep(50 * time.Millisecond)
-	}
-}
-
-// ---- Schemas ----
-
-type MetaFields struct {
-	Mode       string `json:"mode,omitempty"`
-	ModifiedAt string `json:"modified_at,omitempty"`
-}
-
-type ReadArgs struct {
-	Path     string `json:"path"`
-	Encoding string `json:"encoding,omitempty"` // text|base64 (auto if empty)
-	MaxBytes int    `json:"max_bytes,omitempty"`
-}
-
-type ReadResult struct {
-	Path      string `json:"path"`
-	Size      int64  `json:"size"`
-	MIMEType  string `json:"mime_type"`
-	SHA256    string `json:"sha256"`
-	Encoding  string `json:"encoding"`
-	Content   string `json:"content"`
-	Truncated bool   `json:"truncated"`
-	MetaFields
-}
-
-type PeekArgs struct {
-	Path     string `json:"path"`
-	Offset   int    `json:"offset,omitempty"`
-	MaxBytes int    `json:"max_bytes,omitempty"`
-}
-
-type PeekResult struct {
-	Path     string `json:"path"`
-	Offset   int    `json:"offset"`
-	Size     int64  `json:"size"`
-	EOF      bool   `json:"eof"`
-	Encoding string `json:"encoding"`
-	Content  string `json:"content"`
-	MetaFields
-}
-
-type WriteArgs struct {
-	Path       string        `json:"path"`
-	Encoding   string        `json:"encoding"` // text|base64
-	Content    string        `json:"content"`
-	Strategy   writeStrategy `json:"strategy,omitempty"`
-	CreateDirs *bool         `json:"create_dirs,omitempty"`
-	Mode       string        `json:"mode,omitempty"` // e.g. 0644
-	Start      *int          `json:"start,omitempty"`
-	End        *int          `json:"end,omitempty"`
-}
-
-type WriteResult struct {
-	Path     string `json:"path"`
-	Action   string `json:"action"`
-	Bytes    int    `json:"bytes"`
-	Created  bool   `json:"created"`
-	MIMEType string `json:"mime_type"`
-	SHA256   string `json:"sha256"`
-	MetaFields
-}
-
-type EditArgs struct {
-	Path    string `json:"path"`
-	Pattern string `json:"pattern"`
-	Replace string `json:"replace"`
-	Regex   bool   `json:"regex,omitempty"`
-	Count   int    `json:"count,omitempty"`
-}
-
-type EditResult struct {
-	Path         string `json:"path"`
-	Replacements int    `json:"replacements"`
-	Bytes        int    `json:"bytes"`
-	SHA256       string `json:"sha256"`
-	MetaFields
-}
-
-type ListArgs struct {
-	Path       string `json:"path"`
-	Recursive  bool   `json:"recursive,omitempty"`
-	MaxEntries int    `json:"max_entries,omitempty"`
-}
-
-type ListEntry struct {
-	Path       string `json:"path"`
-	Name       string `json:"name"`
-	Kind       string `json:"kind"`
-	Size       int64  `json:"size"`
-	Mode       string `json:"mode"`
-	ModifiedAt string `json:"modified_at"`
-}
-
-type ListResult struct {
-	Entries []ListEntry `json:"entries"`
-}
-
-type GlobArgs struct {
-	Pattern    string `json:"pattern"`
-	MaxResults int    `json:"max_results,omitempty"`
-}
-
-type GlobResult struct {
-	Matches []string `json:"matches"`
-}
-
-type SearchArgs struct {
-	Pattern    string `json:"pattern"`
-	Path       string `json:"path,omitempty"`
-	Regex      bool   `json:"regex,omitempty"`
-	MaxResults int    `json:"max_results,omitempty"`
-}
-
-type SearchMatch struct {
-	Path string `json:"path"`
-	Line int    `json:"line"`
-	Text string `json:"text"`
-}
-
-type SearchResult struct {
-	Matches []SearchMatch `json:"matches"`
-}
-
-func formatReadResult(r ReadResult) string {
-	return fmt.Sprintf("path=%s size=%d mime=%s sha=%s enc=%s truncated=%v content=%s", r.Path, r.Size, r.MIMEType, r.SHA256, r.Encoding, r.Truncated, r.Content)
-}
-
-func formatPeekResult(r PeekResult) string {
-	return fmt.Sprintf("path=%s offset=%d size=%d eof=%v enc=%s content=%s", r.Path, r.Offset, r.Size, r.EOF, r.Encoding, r.Content)
-}
-
-func formatWriteResult(r WriteResult) string {
-	return fmt.Sprintf("path=%s action=%s bytes=%d created=%v mime=%s sha=%s", r.Path, r.Action, r.Bytes, r.Created, r.MIMEType, r.SHA256)
-}
-
-func formatEditResult(r EditResult) string {
-	return fmt.Sprintf("path=%s replacements=%d bytes=%d sha=%s", r.Path, r.Replacements, r.Bytes, r.SHA256)
-}
-
-func formatListResult(r ListResult) string {
-	var b strings.Builder
-	for i, e := range r.Entries {
-		if i > 0 {
-			b.WriteByte('\n')
-		}
-		fmt.Fprintf(&b, "%s %s %s %d %s %s", e.Path, e.Name, e.Kind, e.Size, e.Mode, e.ModifiedAt)
-	}
-	return b.String()
-}
-
-func formatSearchResult(r SearchResult) string {
-	var b strings.Builder
-	for i, m := range r.Matches {
-		if i > 0 {
-			b.WriteByte('\n')
-		}
-		fmt.Fprintf(&b, "%s:%d:%s", m.Path, m.Line, m.Text)
-	}
-	return b.String()
-}
-
-func formatGlobResult(r GlobResult) string {
-	return strings.Join(r.Matches, "\n")
-}
-
-func wrapTextHandler[TArgs any, TResult any](h mcp.StructuredToolHandlerFunc[TArgs, TResult], format func(TResult) string) func(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	return func(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-		var args TArgs
-		if err := req.BindArguments(&args); err != nil {
-			return mcp.NewToolResultError(fmt.Sprintf("failed to bind arguments: %v", err)), nil
-		}
-		res, err := h(ctx, req, args)
-		if err != nil {
-			return mcp.NewToolResultError(fmt.Sprintf("tool execution failed: %v", err)), nil
-		}
-		return mcp.NewToolResultText(format(res)), nil
-	}
-}
-
-func kindOf(fi os.FileInfo) string {
-	m := fi.Mode()
-	if m.IsRegular() {
-		return "file"
-	}
-	if m.IsDir() {
-		return "dir"
-	}
-	if (m & os.ModeSymlink) != 0 {
-		return "symlink"
-	}
-	return "other"
-}
-
-// ---- Handlers ----
-
-func handleRead(root string) mcp.StructuredToolHandlerFunc[ReadArgs, ReadResult] {
-	return func(ctx context.Context, req mcp.CallToolRequest, args ReadArgs) (ReadResult, error) {
-		start := time.Now()
-		dprintf("-> fs_read path=%q encoding=%q max_bytes=%d", args.Path, args.Encoding, args.MaxBytes)
-		var res ReadResult
-		full, err := safeJoinResolveFinal(root, args.Path)
-		if err != nil {
-			dprintf("fs_read error: %v", err)
-			return res, err
-		}
-		fi, err := os.Stat(full)
-		if err != nil {
-			dprintf("fs_read stat error: %v", err)
-			return res, err
-		}
-		limit := args.MaxBytes
-		if limit <= 0 {
-			limit = defaultReadMaxBytes
-		}
-		f, err := os.Open(full)
-		if err != nil {
-			dprintf("fs_read open error: %v", err)
-			return res, err
-		}
-		defer f.Close()
-		r := io.LimitReader(f, int64(limit))
-		buf, err := io.ReadAll(r)
-		if err != nil {
-			dprintf("fs_read read error: %v", err)
-			return res, err
-		}
-		trunc := fi.Size() > int64(len(buf))
-
-		sha := ""
-		if fi.Size() <= maxHashBytes {
-			hf, err := os.Open(full)
-			if err == nil {
-				h := sha256.New()
-				if _, err := io.Copy(h, hf); err == nil {
-					sha = fmt.Sprintf("%x", h.Sum(nil))
-				}
-				hf.Close()
-			}
-		} else {
-			dprintf("fs_read: skip sha256 (size %d > cap %d)", fi.Size(), maxHashBytes)
-		}
-
-		enc := args.Encoding
-		if enc == "" {
-			sample := buf
-			if len(sample) > maxPeekBytesForSniff {
-				sample = sample[:maxPeekBytesForSniff]
-			}
-			if isText(sample) {
-				enc = string(encText)
-			} else {
-				enc = string(encBase64)
-			}
-		}
-		var content string
-		if encodingKind(enc) == encBase64 {
-			content = base64.StdEncoding.EncodeToString(buf)
-		} else {
-			content = string(buf)
-		}
-
-		res = ReadResult{
-			Path:      args.Path,
-			Size:      fi.Size(),
-			MIMEType:  detectMIME(full, buf),
-			SHA256:    sha,
-			Encoding:  enc,
-			Content:   content,
-			Truncated: trunc,
-			MetaFields: MetaFields{
-				Mode:       fmt.Sprintf("%#o", fi.Mode()&os.ModePerm),
-				ModifiedAt: fi.ModTime().UTC().Format(time.RFC3339),
-			},
-		}
-		dprintf("<- fs_read ok size=%d truncated=%v dur=%s", len(buf), trunc, time.Since(start))
-		return res, nil
-	}
-}
-
-func handlePeek(root string) mcp.StructuredToolHandlerFunc[PeekArgs, PeekResult] {
-	return func(ctx context.Context, req mcp.CallToolRequest, args PeekArgs) (PeekResult, error) {
-		start := time.Now()
-		if args.MaxBytes <= 0 {
-			args.MaxBytes = defaultPeekMaxBytes
-		}
-		dprintf("-> fs_peek path=%q offset=%d max_bytes=%d", args.Path, args.Offset, args.MaxBytes)
-		var res PeekResult
-		full, err := safeJoinResolveFinal(root, args.Path)
-		if err != nil {
-			dprintf("fs_peek error: %v", err)
-			return res, err
-		}
-		chunk, sz, eof, err := readWindow(full, args.Offset, args.MaxBytes)
-		if err != nil {
-			dprintf("fs_peek read error: %v", err)
-			return res, err
-		}
-		enc := string(encText)
-		content := string(chunk)
-		if !isText(chunk) {
-			enc = string(encBase64)
-			content = base64.StdEncoding.EncodeToString(chunk)
-		}
-		var mode string
-		var modAt string
-		if fi, statErr := os.Lstat(full); statErr == nil {
-			mode = fmt.Sprintf("%#o", fi.Mode()&os.ModePerm)
-			modAt = fi.ModTime().UTC().Format(time.RFC3339)
-		}
-		res = PeekResult{
-			Path:     args.Path,
-			Offset:   args.Offset,
-			Size:     sz,
-			EOF:      eof,
-			Encoding: enc,
-			Content:  content,
-			MetaFields: MetaFields{
-				Mode:       mode,
-				ModifiedAt: modAt,
-			},
-		}
-		dprintf("<- fs_peek ok bytes=%d eof=%v dur=%s", len(chunk), eof, time.Since(start))
-		return res, nil
-	}
-}
-
-func handleWrite(root string) mcp.StructuredToolHandlerFunc[WriteArgs, WriteResult] {
-	return func(ctx context.Context, req mcp.CallToolRequest, args WriteArgs) (WriteResult, error) {
-		start := time.Now()
-		dprintf("-> fs_write path=%q strategy=%q encoding=%q bytes=%d", args.Path, args.Strategy, args.Encoding, len(args.Content))
-		var res WriteResult
-		if args.Encoding == "" {
-			dprintf("fs_write error: encoding required")
-			return res, errors.New("encoding is required: text|base64")
-		}
-		full, err := safeJoin(root, args.Path)
-		if err != nil {
-			dprintf("fs_write error: %v", err)
-			return res, err
-		}
-		// Do not create parent directories unless explicitly requested.
-		if args.CreateDirs == nil {
-			b := false
-			args.CreateDirs = &b
-		}
-		if *args.CreateDirs {
-			if err := ensureParent(full); err != nil {
-				dprintf("fs_write error: %v", err)
-				return res, err
-			}
-		}
-		mode, err := parseMode(args.Mode)
-		if err != nil {
-			dprintf("fs_write error: %v", err)
-			return res, fmt.Errorf("invalid mode: %w", err)
-		}
-		modeProvided := args.Mode != ""
-		var data []byte
-		if encodingKind(args.Encoding) == encBase64 {
-			b, err := base64.StdEncoding.DecodeString(args.Content)
-			if err != nil {
-				dprintf("fs_write error: %v", err)
-				return res, fmt.Errorf("invalid base64 content: %w", err)
-			}
-			data = b
-		} else {
-			data = []byte(args.Content)
-		}
-		st := args.Strategy
-		if st == "" {
-			st = strategyOverwrite
-		}
-
-		// Pre-stat & symlink/dir guards
-		preFi, preErr := os.Lstat(full)
-		if preErr == nil && (preFi.Mode()&os.ModeSymlink) != 0 {
-			dprintf("fs_write error: target is symlink")
-			return res, fmt.Errorf("refusing to write to symlink: %s", args.Path)
-		}
-		if preErr == nil && preFi.IsDir() && (st == strategyOverwrite || st == strategyNoClobber) {
-			return res, fmt.Errorf("target is a directory: %s", args.Path)
-		}
-		if preErr == nil && !modeProvided {
-			if pm := preFi.Mode() & os.ModePerm; pm != 0 {
-				mode = pm
-			} else {
-				mode = 0o644
-			}
-		}
-
-		release, err := acquireLock(full, 3*time.Second)
-		if err != nil {
-			dprintf("fs_write lock error: %v", err)
-			return res, err
-		}
-		defer release()
-
-		created := false
-		action := string(st)
-
-		switch st {
-		case strategyNoClobber:
-			if preErr == nil {
-				dprintf("fs_write noclobber exists")
-				return res, fmt.Errorf("exists: %s", args.Path)
-			}
-			if err := atomicWrite(full, data, mode); err != nil {
-				dprintf("fs_write error: %v", err)
-				return res, err
-			}
-			created = true
-
-		case strategyOverwrite:
-			if errors.Is(preErr, os.ErrNotExist) {
-				created = true
-			}
-			if err := atomicWrite(full, data, mode); err != nil {
-				dprintf("fs_write error: %v", err)
-				return res, err
-			}
-
-		case strategyAppend:
-			if preErr == nil && !preFi.Mode().IsRegular() {
-				return res, fmt.Errorf("append target not a regular file: %s", args.Path)
-			}
-			if errors.Is(preErr, os.ErrNotExist) {
-				created = true
-			}
-			f, err := os.OpenFile(full, os.O_CREATE|os.O_WRONLY|os.O_APPEND, mode)
-			if err != nil {
-				dprintf("fs_write error: %v", err)
-				return res, err
-			}
-			defer f.Close()
-			n, err := f.Write(data)
-			if err != nil {
-				dprintf("fs_write error: %v", err)
-				return res, err
-			}
-			data = data[:n]
-
-		case strategyPrepend:
-			if preErr == nil && !preFi.Mode().IsRegular() {
-				return res, fmt.Errorf("prepend target not a regular file: %s", args.Path)
-			}
-			var old []byte
-			if preErr == nil {
-				old, err = os.ReadFile(full)
-				if err != nil {
-					return res, err
-				}
-			} else if errors.Is(preErr, os.ErrNotExist) {
-				created = true
-			}
-			buf := append([]byte{}, data...)
-			buf = append(buf, old...)
-			if err := atomicWrite(full, buf, mode); err != nil {
-				dprintf("fs_write error: %v", err)
-				return res, err
-			}
-			data = buf
-
-		case strategyReplaceRange:
-			if preErr != nil {
-				dprintf("fs_write error: %v", preErr)
-				return res, fmt.Errorf("replace_range requires existing file: %w", preErr)
-			}
-			if !preFi.Mode().IsRegular() {
-				return res, fmt.Errorf("replace_range target not a regular file: %s", args.Path)
-			}
-			old, err := os.ReadFile(full)
-			if err != nil {
-				dprintf("fs_write error: %v", err)
-				return res, err
-			}
-			if args.Start == nil || args.End == nil {
-				return res, errors.New("start and end required for replace_range")
-			}
-			s, e := *args.Start, *args.End
-			if s < 0 || e < s || e > len(old) {
-				return res, fmt.Errorf("invalid range [%d,%d)", s, e)
-			}
-			buf := append([]byte{}, old[:s]...)
-			buf = append(buf, data...)
-			buf = append(buf, old[e:]...)
-			if err := atomicWrite(full, buf, mode); err != nil {
-				dprintf("fs_write error: %v", err)
-				return res, err
-			}
-			data = buf
-
-		default:
-			return res, fmt.Errorf("unknown strategy: %s", st)
-		}
-
-		final := data
-		if b, err := os.ReadFile(full); err == nil {
-			final = b
-		}
-		mt := detectMIME(full, final)
-		fi, statErr := os.Lstat(full)
-		modAt := time.Now().UTC().Format(time.RFC3339)
-		modeStr := ""
-		if fi != nil && statErr == nil {
-			modAt = fi.ModTime().UTC().Format(time.RFC3339)
-			modeStr = fmt.Sprintf("%#o", fi.Mode()&os.ModePerm)
-		}
-		sha := ""
-		if len(final) <= int(maxHashBytes) {
-			sha = sha256sum(final)
-		} else {
-			dprintf("fs_write: skip sha256 (size %d > cap %d)", len(final), maxHashBytes)
-		}
-		res = WriteResult{
-			Path:     args.Path,
-			Action:   action,
-			Bytes:    len(final),
-			Created:  created,
-			MIMEType: mt,
-			SHA256:   sha,
-			MetaFields: MetaFields{
-				Mode:       modeStr,
-				ModifiedAt: modAt,
-			},
-		}
-		dprintf("<- fs_write ok created=%v bytes=%d dur=%s", created, len(final), time.Since(start))
-		return res, nil
-	}
-}
-
-func handleEdit(root string) mcp.StructuredToolHandlerFunc[EditArgs, EditResult] {
-	return func(ctx context.Context, req mcp.CallToolRequest, args EditArgs) (EditResult, error) {
-		start := time.Now()
-		dprintf("-> fs_edit path=%q regex=%v count=%d", args.Path, args.Regex, args.Count)
-		var res EditResult
-		if args.Path == "" || args.Pattern == "" {
-			return res, errors.New("path and pattern required")
-		}
-		full, err := safeJoin(root, args.Path)
-		if err != nil {
-			dprintf("fs_edit error: %v", err)
-			return res, err
-		}
-		fi, err := os.Lstat(full)
-		if err != nil {
-			dprintf("fs_edit error: %v", err)
-			return res, err
-		}
-		if (fi.Mode() & os.ModeSymlink) != 0 {
-			return res, fmt.Errorf("refusing to edit symlink: %s", args.Path)
-		}
-		if !fi.Mode().IsRegular() {
-			return res, fmt.Errorf("target not a regular file: %s", args.Path)
-		}
-
-		release, err := acquireLock(full, 3*time.Second)
-		if err != nil {
-			dprintf("fs_edit lock error: %v", err)
-			return res, err
-		}
-		defer release()
-
-		b, err := os.ReadFile(full)
-		if err != nil {
-			dprintf("fs_edit read error: %v", err)
-			return res, err
-		}
-		var re *regexp.Regexp
-		if args.Regex {
-			re, err = regexp.Compile(args.Pattern)
-			if err != nil {
-				return res, fmt.Errorf("invalid regex: %w", err)
-			}
-		}
-		count := 0
-		var out []byte
-		if args.Regex {
-			if args.Count <= 0 {
-				out = re.ReplaceAll(b, []byte(args.Replace))
-				count = len(re.FindAllIndex(b, -1))
-			} else {
-				remaining := args.Count
-				out = re.ReplaceAllFunc(b, func(m []byte) []byte {
-					if remaining == 0 {
-						return m
-					}
-					remaining--
-					count++
-					return []byte(args.Replace)
-				})
-			}
-		} else {
-			old := string(b)
-			limit := args.Count
-			if limit <= 0 {
-				out = []byte(strings.ReplaceAll(old, args.Pattern, args.Replace))
-				if args.Pattern != "" {
-					count = strings.Count(old, args.Pattern)
-				}
-			} else {
-				out = []byte(strings.Replace(old, args.Pattern, args.Replace, limit))
-				if args.Pattern != "" {
-					if c := strings.Count(old, args.Pattern); c < limit {
-						count = c
-					} else {
-						count = limit
-					}
-				}
-			}
-		}
-		mode := fi.Mode() & os.ModePerm
-		if mode == 0 {
-			mode = 0o644
-		}
-		if err := atomicWrite(full, out, mode); err != nil {
-			dprintf("fs_edit write error: %v", err)
-			return res, err
-		}
-		res = EditResult{
-			Path:         args.Path,
-			Replacements: count,
-			Bytes:        len(out),
-			SHA256:       sha256sum(out),
-			MetaFields: MetaFields{
-				Mode:       fmt.Sprintf("%#o", mode),
-				ModifiedAt: time.Now().UTC().Format(time.RFC3339),
-			},
-		}
-		dprintf("<- fs_edit ok replacements=%d bytes=%d dur=%s", count, len(out), time.Since(start))
-		return res, nil
-	}
-}
-
-func handleList(root string) mcp.StructuredToolHandlerFunc[ListArgs, ListResult] {
-	return func(ctx context.Context, req mcp.CallToolRequest, args ListArgs) (ListResult, error) {
-		start := time.Now()
-		dprintf("-> fs_list path=%q recursive=%v max_entries=%d", args.Path, args.Recursive, args.MaxEntries)
-		var out ListResult
-		base, err := safeJoinResolveFinal(root, args.Path)
-		if err != nil {
-			dprintf("fs_list error: %v", err)
-			return out, err
-		}
-		max := args.MaxEntries
-		if max <= 0 {
-			max = defaultListMaxEntries
-		}
-		count := 0
-		add := func(path string, fi os.FileInfo) {
-			if count >= max {
-				return
-			}
-			out.Entries = append(out.Entries, ListEntry{
-				Path:       filepath.ToSlash(trimUnderRoot(root, path)),
-				Name:       fi.Name(),
-				Kind:       kindOf(fi),
-				Size:       fi.Size(),
-				Mode:       fmt.Sprintf("%#o", fi.Mode()&os.ModePerm),
-				ModifiedAt: fi.ModTime().UTC().Format(time.RFC3339),
-			})
-			count++
-		}
-		fi, err := os.Stat(base)
-		if err != nil {
-			dprintf("fs_list stat error: %v", err)
-			return out, err
-		}
-		if fi.IsDir() {
-			if !args.Recursive {
-				ents, err := os.ReadDir(base)
-				if err != nil {
-					dprintf("fs_list readdir error: %v", err)
-					return out, err
-				}
-				for _, e := range ents {
-					select {
-					case <-ctx.Done():
-						return out, ctx.Err()
-					default:
-					}
-					info, err := e.Info()
-					if err != nil {
-						continue
-					}
-					add(filepath.Join(base, e.Name()), info)
-				}
-			} else {
-				err := filepath.Walk(base, func(path string, info os.FileInfo, err error) error {
-					if err != nil {
-						return nil
-					}
-					select {
-					case <-ctx.Done():
-						return ctx.Err()
-					default:
-					}
-					add(path, info)
-					if count >= max {
-						return io.EOF
-					}
-					return nil
-				})
-				if err != nil && !errors.Is(err, io.EOF) {
-					dprintf("fs_list walk error: %v", err)
-					return out, err
-				}
-			}
-		} else {
-			add(base, fi)
-		}
-		dprintf("<- fs_list ok entries=%d dur=%s", len(out.Entries), time.Since(start))
-		return out, nil
-	}
-}
-
-func handleSearch(root string) mcp.StructuredToolHandlerFunc[SearchArgs, SearchResult] {
-	return func(ctx context.Context, req mcp.CallToolRequest, args SearchArgs) (SearchResult, error) {
-		start := time.Now()
-		dprintf("-> fs_search path=%q pattern=%q regex=%v max=%d", args.Path, args.Pattern, args.Regex, args.MaxResults)
-		var out SearchResult
-		if args.Pattern == "" {
-			return out, errors.New("pattern required")
-		}
-		max := args.MaxResults
-		if max <= 0 {
-			max = defaultSearchMaxResults
-		}
-		var rx *regexp.Regexp
-		if args.Regex {
-			r, err := regexp.Compile(args.Pattern)
-			if err != nil {
-				dprintf("fs_search error: %v", err)
-				return out, err
-			}
-			rx = r
-		}
-		startPath := root
-		if args.Path != "" {
-			p, err := safeJoin(root, args.Path)
-			if err != nil {
-				dprintf("fs_search error: %v", err)
-				return out, err
-			}
-			startPath = p
-		}
-		if _, err := os.Stat(startPath); err != nil {
-			dprintf("fs_search error: %v", err)
-			return out, err
-		}
-		ctx, cancel := context.WithCancel(ctx)
-		defer cancel()
-
-		files := make(chan string, 64)
-		var walkErr error
-		var walkWG sync.WaitGroup
-		walkWG.Add(1)
-		go func() {
-			defer walkWG.Done()
-			walkErr = filepath.WalkDir(startPath, func(path string, d fs.DirEntry, err error) error {
-				if err != nil {
-					return nil
-				}
-				select {
-				case <-ctx.Done():
-					return ctx.Err()
-				default:
-				}
-				if d.IsDir() {
-					return nil
-				}
-				if d.Type()&os.ModeSymlink != 0 {
-					return nil
-				}
-				files <- path
-				return nil
-			})
-			close(files)
-		}()
-
-		var mu sync.Mutex
-		matches := []SearchMatch{}
-		workers := runtime.NumCPU()
-		var wg sync.WaitGroup
-		for i := 0; i < workers; i++ {
-			wg.Add(1)
-			go func() {
-				defer wg.Done()
-				for path := range files {
-					if ctx.Err() != nil {
-						return
-					}
-					f, err := os.Open(path)
-					if err != nil {
-						continue
-					}
-					scanner := bufio.NewScanner(f)
-					scanner.Buffer(make([]byte, 64*1024), 4*1024*1024)
-					lineNo := 1
-					for scanner.Scan() {
-						txt := scanner.Text()
-						var ok bool
-						if rx != nil {
-							ok = rx.MatchString(txt)
-						} else {
-							ok = strings.Contains(txt, args.Pattern)
-						}
-						if ok {
-							rel, _ := filepath.Rel(root, path)
-							mu.Lock()
-							matches = append(matches, SearchMatch{Path: filepath.ToSlash(rel), Line: lineNo, Text: txt})
-							if len(matches) >= max {
-								mu.Unlock()
-								cancel()
-								f.Close()
-								return
-							}
-							mu.Unlock()
-						}
-						lineNo++
-					}
-					f.Close()
-					if err := scanner.Err(); err != nil && !errors.Is(err, io.EOF) {
-						continue
-					}
-				}
-			}()
-		}
-		wg.Wait()
-		walkWG.Wait()
-		if walkErr != nil && !errors.Is(walkErr, context.Canceled) {
-			dprintf("fs_search error: %v", walkErr)
-			return out, walkErr
-		}
-		out.Matches = matches
-		dprintf("<- fs_search ok matches=%d dur=%s", len(out.Matches), time.Since(start))
-		return out, nil
-	}
-}
-
-func handleGlob(root string) mcp.StructuredToolHandlerFunc[GlobArgs, GlobResult] {
-	return func(ctx context.Context, req mcp.CallToolRequest, args GlobArgs) (GlobResult, error) {
-		start := time.Now()
-		dprintf("-> fs_glob pattern=%q max_results=%d", args.Pattern, args.MaxResults)
-		var out GlobResult
-		if args.Pattern == "" {
-			return out, errors.New("pattern required")
-		}
-		if _, err := safeJoin(root, args.Pattern); err != nil {
-			dprintf("fs_glob error: %v", err)
-			return out, err
-		}
-		max := args.MaxResults
-		if max <= 0 {
-			max = defaultGlobMaxResults
-		}
-		pat := filepath.ToSlash(filepath.Clean(args.Pattern))
-		if _, err := doublestar.Match(pat, ""); err != nil {
-			dprintf("fs_glob error: %v", err)
-			return out, err
-		}
-		ctx, cancel := context.WithCancel(ctx)
-		defer cancel()
-		paths := make(chan string, 64)
-		var walkErr error
-		var walkWG sync.WaitGroup
-		walkWG.Add(1)
-		go func() {
-			defer walkWG.Done()
-			walkErr = filepath.WalkDir(root, func(path string, d fs.DirEntry, err error) error {
-				if err != nil {
-					return nil
-				}
-				select {
-				case <-ctx.Done():
-					return ctx.Err()
-				default:
-				}
-				rel, err := filepath.Rel(root, path)
-				if err != nil {
-					return nil
-				}
-				paths <- filepath.ToSlash(rel)
-				return nil
-			})
-			close(paths)
-		}()
-
-		var mu sync.Mutex
-		matches := []string{}
-		workers := runtime.NumCPU()
-		var wg sync.WaitGroup
-		for i := 0; i < workers; i++ {
-			wg.Add(1)
-			go func() {
-				defer wg.Done()
-				for p := range paths {
-					if ctx.Err() != nil {
-						return
-					}
-					ok, err := doublestar.Match(pat, p)
-					if err != nil {
-						cancel()
-						return
-					}
-					if ok {
-						mu.Lock()
-						if len(matches) >= max {
-							mu.Unlock()
-							return
-						}
-						matches = append(matches, filepath.ToSlash(p))
-						if len(matches) >= max {
-							mu.Unlock()
-							cancel()
-							return
-						}
-						mu.Unlock()
-					}
-				}
-			}()
-		}
-		wg.Wait()
-		walkWG.Wait()
-		if walkErr != nil && !errors.Is(walkErr, context.Canceled) {
-			dprintf("fs_glob error: %v", walkErr)
-			return out, walkErr
-		}
-		out.Matches = matches
-		dprintf("<- fs_glob ok matches=%d dur=%s", len(out.Matches), time.Since(start))
-		return out, nil
-	}
-}
-
-// ---- main ----
 
 func main() {
 	flag.Parse()
@@ -1143,127 +22,7 @@ func main() {
 	}
 	dprintf("server start root=%q debug=%v", root, debugEnabled)
 
-	s := server.NewMCPServer("fs-mcp-go", "0.1.0")
-
-	readOpts := []mcp.ToolOption{
-		mcp.WithDescription("Read a file up to a byte limit. Detects encoding when unspecified."),
-		mcp.WithString("path", mcp.Required(), mcp.Description("File path or file:// URI within root")),
-		mcp.WithString("encoding", mcp.Enum(string(encText), string(encBase64)), mcp.Description("Force text or base64; auto-detected if empty")),
-		mcp.WithNumber("max_bytes", mcp.Min(1), mcp.Description("Maximum bytes to return (default 64 KiB)")),
-	}
-	if !*compatFlag {
-		readOpts = append(readOpts, mcp.WithOutputSchema[ReadResult]())
-	}
-	readTool := mcp.NewTool("fs_read", readOpts...)
-	if *compatFlag {
-		s.AddTool(readTool, wrapTextHandler(handleRead(root), formatReadResult))
-	} else {
-		s.AddTool(readTool, mcp.NewStructuredToolHandler(handleRead(root)))
-	}
-
-	peekOpts := []mcp.ToolOption{
-		mcp.WithDescription("Read a file window without loading the whole file"),
-		mcp.WithString("path", mcp.Required(), mcp.Description("File path")),
-		mcp.WithNumber("offset", mcp.Min(0), mcp.Description("Byte offset to start at (default 0)")),
-		mcp.WithNumber("max_bytes", mcp.Min(1), mcp.Description("Window size in bytes (default 4 KiB)")),
-	}
-	if !*compatFlag {
-		peekOpts = append(peekOpts, mcp.WithOutputSchema[PeekResult]())
-	}
-	peekTool := mcp.NewTool("fs_peek", peekOpts...)
-	if *compatFlag {
-		s.AddTool(peekTool, wrapTextHandler(handlePeek(root), formatPeekResult))
-	} else {
-		s.AddTool(peekTool, mcp.NewStructuredToolHandler(handlePeek(root)))
-	}
-
-	writeOpts := []mcp.ToolOption{
-		mcp.WithDescription("Create or modify a file using a strategy"),
-		mcp.WithString("path", mcp.Required(), mcp.Description("Target file path")),
-		mcp.WithString("encoding", mcp.Required(), mcp.Enum(string(encText), string(encBase64)), mcp.Description("Content encoding: text or base64")),
-		mcp.WithString("content", mcp.Required(), mcp.Description("Data to write")),
-		mcp.WithString("strategy", mcp.Enum(string(strategyOverwrite), string(strategyNoClobber), string(strategyAppend), string(strategyPrepend), string(strategyReplaceRange)), mcp.Description("Write behavior (default overwrite)")),
-		mcp.WithBoolean("create_dirs", mcp.Description("Create parent directories if needed (default false)")),
-		mcp.WithString("mode", mcp.Pattern("^0?[0-7]{3,4}$"), mcp.Description("File mode in octal; omit to keep existing")),
-		mcp.WithNumber("start", mcp.Min(0), mcp.Description("Start byte for replace_range")),
-		mcp.WithNumber("end", mcp.Min(0), mcp.Description("End byte (exclusive) for replace_range")),
-	}
-	if !*compatFlag {
-		writeOpts = append(writeOpts, mcp.WithOutputSchema[WriteResult]())
-	}
-	writeTool := mcp.NewTool("fs_write", writeOpts...)
-	if *compatFlag {
-		s.AddTool(writeTool, wrapTextHandler(handleWrite(root), formatWriteResult))
-	} else {
-		s.AddTool(writeTool, mcp.NewStructuredToolHandler(handleWrite(root)))
-	}
-
-	editOpts := []mcp.ToolOption{
-		mcp.WithDescription("Search and replace text in a file"),
-		mcp.WithString("path", mcp.Required(), mcp.Description("Target text file")),
-		mcp.WithString("pattern", mcp.Required(), mcp.Description("Substring or regex to match")),
-		mcp.WithString("replace", mcp.Required(), mcp.Description("Replacement text; supports $1 etc. in regex mode")),
-		mcp.WithBoolean("regex", mcp.Description("Treat pattern as a regular expression")),
-		mcp.WithNumber("count", mcp.Min(0), mcp.Description("If >0, maximum replacements; 0 replaces all")),
-	}
-	if !*compatFlag {
-		editOpts = append(editOpts, mcp.WithOutputSchema[EditResult]())
-	}
-	editTool := mcp.NewTool("fs_edit", editOpts...)
-	if *compatFlag {
-		s.AddTool(editTool, wrapTextHandler(handleEdit(root), formatEditResult))
-	} else {
-		s.AddTool(editTool, mcp.NewStructuredToolHandler(handleEdit(root)))
-	}
-
-	listOpts := []mcp.ToolOption{
-		mcp.WithDescription("List directory contents"),
-		mcp.WithString("path", mcp.Required(), mcp.Description("Directory to list")),
-		mcp.WithBoolean("recursive", mcp.Description("Recurse into subdirectories")),
-		mcp.WithNumber("max_entries", mcp.Min(1), mcp.Description("Maximum entries to return (default 1000)")),
-	}
-	if !*compatFlag {
-		listOpts = append(listOpts, mcp.WithOutputSchema[ListResult]())
-	}
-	listTool := mcp.NewTool("fs_list", listOpts...)
-	if *compatFlag {
-		s.AddTool(listTool, wrapTextHandler(handleList(root), formatListResult))
-	} else {
-		s.AddTool(listTool, mcp.NewStructuredToolHandler(handleList(root)))
-	}
-
-	searchOpts := []mcp.ToolOption{
-		mcp.WithDescription("Search files recursively for text using concurrent workers"),
-		mcp.WithString("pattern", mcp.Required(), mcp.Description("Substring or regex to find")),
-		mcp.WithString("path", mcp.Description("Start directory relative to root")),
-		mcp.WithBoolean("regex", mcp.Description("Interpret pattern as regular expression")),
-		mcp.WithNumber("max_results", mcp.Min(1), mcp.Description("Maximum matches to return (default 100)")),
-	}
-	if !*compatFlag {
-		searchOpts = append(searchOpts, mcp.WithOutputSchema[SearchResult]())
-	}
-	searchTool := mcp.NewTool("fs_search", searchOpts...)
-	if *compatFlag {
-		s.AddTool(searchTool, wrapTextHandler(handleSearch(root), formatSearchResult))
-	} else {
-		s.AddTool(searchTool, mcp.NewStructuredToolHandler(handleSearch(root)))
-	}
-
-	globOpts := []mcp.ToolOption{
-		mcp.WithDescription("Match paths with shell-style globbing and ** for recursion"),
-		mcp.WithString("pattern", mcp.Required(), mcp.Description("Glob pattern relative to root")),
-		mcp.WithNumber("max_results", mcp.Min(1), mcp.Description("Maximum matches to return (default 1000)")),
-	}
-	if !*compatFlag {
-		globOpts = append(globOpts, mcp.WithOutputSchema[GlobResult]())
-	}
-	globTool := mcp.NewTool("fs_glob", globOpts...)
-	if *compatFlag {
-		s.AddTool(globTool, wrapTextHandler(handleGlob(root), formatGlobResult))
-	} else {
-		s.AddTool(globTool, mcp.NewStructuredToolHandler(handleGlob(root)))
-	}
-
+	s := setupServer(root)
 	if err := server.ServeStdio(s); err != nil {
 		_, _ = fmt.Fprintf(os.Stderr, "server error: %v\n", err)
 		dprintf("server error: %v", err)
